delete finra helper. move all functions to general.py
